move printNode to utils
